refactor: freetype-py code organization
